COMMERCE-1217 - portal-vulcan-impl - Move logic from JSONMessageReader/Writer to base classes
