+++ modules/apps/portal-vulcan/portal-vulcan-impl/src/main/java/com/liferay/portal/vulcan/internal/jaxrs/message/body/JSONMessageBodyReader.java
@@ -15,117 +15,21 @@
 package com.liferay.portal.vulcan.internal.jaxrs.message.body;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.ObjectReader;
-
-import com.liferay.petra.string.StringBundler;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.vulcan.internal.jaxrs.validation.ValidatorFactory;
-
-import java.io.IOException;
-import java.io.InputStream;
-
-import java.lang.annotation.Annotation;
-import java.lang.reflect.Type;
-
-import java.util.Optional;
-import java.util.Set;
-
-import javax.servlet.http.HttpServletRequest;
-
-import javax.validation.ConstraintViolation;
-import javax.validation.ValidationException;
-import javax.validation.Validator;
 
 import javax.ws.rs.Consumes;
-import javax.ws.rs.HttpMethod;
-import javax.ws.rs.InternalServerErrorException;
-import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.MultivaluedMap;
-import javax.ws.rs.ext.MessageBodyReader;
 import javax.ws.rs.ext.Provider;
-import javax.ws.rs.ext.Providers;
 
 /**
  * @author Javier Gamarra
+ * @author Ivica Cardic
  */
-@Consumes("application/json")
+@Consumes(MediaType.APPLICATION_JSON)
 @Provider
-public class JSONMessageBodyReader implements MessageBodyReader {
+public class JSONMessageBodyReader extends BaseMessageBodyReader {
 
-	@Override
-	public boolean isReadable(
-		Class clazz, Type type, Annotation[] annotations, MediaType mediaType) {
-
-		return true;
+	public JSONMessageBodyReader() {
+		super(ObjectMapper.class, MediaType.APPLICATION_JSON_TYPE);
 	}
-
-	@Override
-	public Object readFrom(
-			Class clazz, Type type, Annotation[] annotations,
-			MediaType mediaType, MultivaluedMap multivaluedMap,
-			InputStream inputStream)
-		throws IOException {
-
-		ObjectReader objectMapper = _getObjectMapper(
-			clazz
-		).readerFor(
-			clazz
-		);
-
-		Object value = objectMapper.readValue(inputStream);
-
-		if (!StringUtil.equals(
-				_httpServletRequest.getMethod(), HttpMethod.PATCH)) {
-
-			_validate(value);
-		}
-
-		return value;
-	}
-
-	private ObjectMapper _getObjectMapper(Class<?> clazz) {
-		return Optional.ofNullable(
-			_providers.getContextResolver(
-				ObjectMapper.class, MediaType.APPLICATION_JSON_TYPE)
-		).map(
-			contextResolver -> contextResolver.getContext(clazz)
-		).map(
-			ObjectMapper::copy
-		).orElseThrow(
-			() -> new InternalServerErrorException(
-				"Unable to generate object mapper for class " + clazz)
-		);
-	}
-
-	private void _validate(Object value) {
-		Validator validator = ValidatorFactory.getValidator();
-
-		Set<ConstraintViolation<Object>> constraintViolations =
-			validator.validate(value);
-
-		if (constraintViolations.isEmpty()) {
-			return;
-		}
-
-		StringBundler sb = new StringBundler(constraintViolations.size() * 3);
-
-		for (ConstraintViolation<Object> constraintViolation :
-				constraintViolations) {
-
-			sb.append(constraintViolation.getPropertyPath());
-			sb.append(StringPool.SPACE);
-			sb.append(constraintViolation.getMessage());
-		}
-
-		throw new ValidationException(sb.toString());
-	}
-
-	@Context
-	private HttpServletRequest _httpServletRequest;
-
-	@Context
-	private Providers _providers;
 
 }
